requirements.txt
djangorestframework==3.16.0
django-bootstrap5==25.1
django-bootstrap-datepicker-plus==5.0.5
django-dump-load-utf8==0.0.4
django-extensions==4.1
gunicorn==23.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pylint==3.0.2